fix dynamic props in playground
