+++ website/docs/gallery/autorun.md
@@ -4,7 +4,7 @@ hide_table_of_contents: true
 ---
 
 import RacingBars from '../../src/components/RacingBars';
-import {autorunTrueOptions, autorunFalseOptions} from './gallery-demos.ts';
+import { autorunTrueOptions, autorunFalseOptions } from './\_gallery-demos.ts';
 
 A demo for using [`autorun`](../documentation/options.md#autorun) to control auto-starting the chart race.
 
@@ -16,7 +16,7 @@ A demo for using [`autorun`](../documentation/options.md#autorun) to control aut
 
 <div className="gallery">
   <RacingBars
-    {...autorunTrueOptions}
+    {...autorunTrue}
   />
 </div>
 
@@ -26,6 +26,6 @@ A demo for using [`autorun`](../documentation/options.md#autorun) to control aut
 
 <div className="gallery">
   <RacingBars
-    {...autorunFalseOptions}
+    {...autorunFalse}
   />
 </div>
