+++ website/docs/gallery/caption-data-function.md
@@ -4,6 +4,7 @@ hide_table_of_contents: true
 ---
 
 import RacingBars from '../../src/components/RacingBars';
+import { captionDataFunction } from './\_gallery-demos.ts';
 
 This example shows the use of [data function](../documentation/options#data-function) in [caption](../documentation/options#caption).
 
@@ -11,17 +12,8 @@ This example shows the use of [data function](../documentation/options#data-func
 
 ### Chart
 
-export const options = {
-caption: (currentDate, dateSlice, allDates) =>
-`Total: ${Math.round(dateSlice.reduce((acc, curr) => acc + curr.value, 0))}`,
-};
-
 <div className="gallery">
   <RacingBars
-    dataUrl="/data/population.csv"
-    dataType="csv"
-    caption={options.caption}
-    dynamicProps={{caption: `(currentDate, dateSlice, allDates) =>
-\`Total: \${Math.round(dateSlice.reduce((acc, curr) => acc + curr.value, 0))}\``}}
+    {...captionDataFunction}
   />
 </div>
